feat: Update AWS instance resource in main.tf
The code changes modify the count of the AWS instance resource in the main.tf file. The count is changed from 2 to 1. This change reduces the number of instances created. The tags for identifying the instance are also updated to include the count index.

Recent user commits:
- feat: Add AWS instance resource to main.tf
- refactor the file-layout

Recent repository commits:
- chore: Update .gitignore to include .terraform.lock.hcl file
- chore: Add .gitignore and main.tf files