feat: Update AMI validation in variables.tf
The code changes in variables.tf include adding a validation block for the "ami" variable. The validation condition checks if the length of the AMI value is greater than 4 and if it starts with "ami-". If the condition is not met, an error message is displayed. This update improves the validation of the AMI value.

Recent user commits:
- feat: Update AWS instance resource in main.tf
- refactor the file-layout

Recent repository commits:
- chore: Update .gitignore to include .terraform.lock.hcl file
- chore: Add .gitignore and main.tf files