refactor: Update aws_security_group to use default_tags from common_tags variable
The code changes in providers.tf update the aws_security_group resource to use the default_tags from the common_tags variable. This allows for consistent tagging of the security group with the specified common tags. This refactor improves the maintainability and consistency of the Terraform configuration.

Recent user commits:
- feat: Refactor ingress rules in aws_security_group to use dynamic configuration
- feat: Add dynamic ingress rules for specified ports in aws_security_group
- feat: Add server_build variable for database server creation
- refactor: Update AWS instance resource count and tags in main.tf
- added a new vpc
- feat: Add AWS subnets for public availability zones
- feat: Add AWS VPC resource with DNS hostnames enabled

Recent repository commits:
- chore: Update .gitignore to include .terraform.lock.hcl file
- chore: Add .gitignore and main.tf files