feat: Add S3 bucket resource and bucket policy
The code changes in variables.tf, s3.tf, and bucket_policy.tpl add a new S3 bucket resource and bucket policy. The bucket is created with the name specified in the "s3bucket_name" variable, and the bucket policy allows public access to all objects in the bucket. This update enhances the infrastructure by providing a new S3 bucket for storing and accessing files.